+++ resources/js/components/link.tsx
@@ -1,10 +1,10 @@
 import { Button } from '@/components/ui/button';
+import { useIcon } from '@/lib/icons';
 import { cn } from '@/lib/utils';
 import { type Link } from '@/types/Link';
-import * as LucideIcons from 'lucide-react';
 
 export function Link({ ...props }: Link) {
-    const Icon = props.icon ? LucideIcons[props.icon] : null;
+    const Icon = useIcon(props.icon);
 
     return (
         <Button variant={props.variant} className={cn(props.variant === 'link' ? 'px-0' : undefined, props.className)}>