+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+- We may present Image URL's, preserving them in the content root directory.
 - Try deploy hook.
 
 ## [0.3.2] - 2026-02-27